+++ src/options.js
@@ -1,5 +1,5 @@
 import {parse as isoParse} from "isoformat";
-import {color, descending, quantile} from "d3";
+import {color, descending, range as rangei, quantile} from "d3";
 
 // https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/TypedArray
 const TypedArray = Object.getPrototypeOf(Uint8Array);
@@ -233,6 +233,24 @@ export function mid(x1, x2) {
     },
     label: x1.label
   };
+}
+
+// TODO Allow the interval to be specified as a string, e.g. “day” or “hour”?
+// This will require the interval knowing the type of the associated scale to
+// chose between UTC and local time (or better, an explicit timeZone option).
+export function maybeInterval(interval) {
+  if (interval == null) return;
+  if (typeof interval === "number") {
+    const n = interval;
+    return {
+      floor: (d) => n * Math.floor(d / n),
+      offset: (d) => d + n, // note: no optional step for simplicity
+      range: (lo, hi) => rangei(Math.ceil(lo / n), hi / n).map((x) => n * x)
+    };
+  }
+  if (typeof interval.floor !== "function") throw new Error("invalid interval; missing floor method");
+  if (typeof interval.offset !== "function") throw new Error("invalid interval; missing offset method");
+  return interval;
 }
 
 // This distinguishes between per-dimension options and a standalone value.
