+++ src/scales/ordinal.js
@@ -1,8 +1,7 @@
 import {InternSet, extent, quantize, reverse as reverseof, sort, symbolsFill, symbolsStroke} from "d3";
 import {scaleBand, scaleOrdinal, scalePoint, scaleImplicit} from "d3";
 import {ascendingDefined} from "../defined.js";
-import {isNoneish, map} from "../options.js";
-import {maybeInterval} from "../transforms/interval.js";
+import {isNoneish, map, maybeInterval} from "../options.js";
 import {maybeSymbol} from "../symbols.js";
 import {registry, color, position, symbol} from "./index.js";
 import {maybeBooleanRange, ordinalScheme, quantitativeScheme} from "./schemes.js";
