+++ src/scales/quantitative.js
@@ -24,9 +24,8 @@ import {
   ticks
 } from "d3";
 import {positive, negative, finite} from "../defined.js";
-import {arrayify, constant, order, slice} from "../options.js";
+import {arrayify, constant, order, slice, maybeInterval} from "../options.js";
 import {ordinalRange, quantitativeScheme} from "./schemes.js";
-import {maybeInterval} from "../transforms/interval.js";
 import {registry, radius, opacity, color, length} from "./index.js";
 
 export const flip = (i) => (t) => i(1 - t);
